Trying to fix some bugs
