+++ app/controllers/events_controller.rb
@@ -47,4 +47,13 @@ class EventsController < ApplicationController
     @event.delete
   end
 
+  private
+    def event_params
+      params.require(:event).permit(
+        :title,
+        :description,
+        :event_date,
+        :location
+      )
+    end
 end
