+++ app/helpers/events_helper.rb
@@ -1,13 +1,4 @@
 module EventsHelper
-  def event_params
-    params.require(:event).permit(
-      :title,
-      :description,
-      :event_date,
-      :location
-    )
-  end
-
   def upcoming?
     Event.upcoming.include?(self)
   end
