+++ app/models/user.rb
@@ -11,7 +11,7 @@ class User < ApplicationRecord
                     format: { with: VALID_EMAIL_REGEX },
                     uniqueness: { case_sensitive: false }
 
- has_many :attendances, foreign_key: :event_attendee_id  # event_creator_id
+ has_many :attendances, foreign_key: :event_attendee_id
  has_many :attended_events, through: :attendances
  has_many :created_events, dependent: :destroy, foreign_key: :creator_id, class_name: 'Event'
